+++ mybook/book/admin.py
@@ -1,43 +1,7 @@
 from django.contrib import admin
 from .models import Book, User
-from django import forms
 from django.contrib.auth.models import Group
-
-
-class UserCreationForm(forms.ModelForm):
-    password = forms.CharField(label='password', widget=forms.PasswordInput)
-    password_confirm = forms.CharField(label='password_confirm', widget=forms.PasswordInput)
-
-    class Meta:
-        model = User
-        fields = ('email', 'nick_name')
-
-    def clean_password_confirm(self):
-        password = self.cleaned_data.get('password')
-        password_confirm = self.cleaned_data.get('password_confirm')
-
-        if password and password_confirm and password != password_confirm:
-            raise forms.ValidationError("password dont match")
-
-        return password
-
-    def save(self, commit=True):
-        user = super().save(commit=False)
-        user.set_password(self.cleaned_data['password'])
-
-        if commit:
-            user.save()
-
-        return user
-
-
-class UserChangeForm(forms.ModelForm):
-    class Meta:
-        model = User
-        fields = ('email', 'nick_name')
-
-    def clean_password(self):
-        return self.cleaned_data['password']
+from .forms import UserChangeForm, UserCreationForm
 
 
 class BookAdmin(admin.ModelAdmin):
